修改下载信息显示,修改下载功能错误,升级hilt-work版本到最新版

+++ app/src/main/java/cn/xybbz/config/download/core/IDownloadDispatcher.kt
@@ -6,7 +6,7 @@ interface IDownloadDispatcher {
 
     fun onProgress(
         downloadId: Long,
-        progress: Int,
+        progress: Float,
         downloadedBytes: Long,
         totalBytes: Long,
         speedBps: Long
